refactor: consolidate position and move insertion methods in database managers
